LPS-194364 Move LayoutCopyHelper to helper package
