+++ modules/apps/layout/layout-api/src/main/java/com/liferay/layout/helper/LayoutCopyHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.util;
+package com.liferay.layout.helper;
 
 import com.liferay.portal.kernel.model.Layout;
