+++ modules/apps/layout/layout-api/src/main/resources/com/liferay/layout/helper/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
